move API tests (#140993) (#141056)
