+++ src/vs/workbench/api/test/browser/extHostDecorations.test.ts
