+++ src/vs/workbench/api/test/browser/extHostDiagnostics.test.ts
