+++ src/vs/workbench/api/test/browser/extHostDocumentData.test.perf-data.ts
