+++ src/vs/workbench/api/test/browser/extHostTesting.test.ts
