+++ src/vs/workbench/api/test/browser/extHostTextEditor.test.ts
