+++ src/vs/workbench/api/test/browser/extHostTypes.test.ts
