+++ src/vs/workbench/api/test/browser/mainThreadConfiguration.test.ts
@@ -11,7 +11,7 @@ import { Extensions, IConfigurationRegistry, ConfigurationScope } from 'vs/platf
 import { IWorkspaceContextService, WorkbenchState } from 'vs/platform/workspace/common/workspace';
 import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
 import { MainThreadConfiguration } from 'vs/workbench/api/browser/mainThreadConfiguration';
-import { SingleProxyRPCProtocol } from './testRPCProtocol';
+import { SingleProxyRPCProtocol } from 'vs/workbench/api/test/common/testRPCProtocol';
 import { IConfigurationService, ConfigurationTarget } from 'vs/platform/configuration/common/configuration';
 import { WorkspaceService } from 'vs/workbench/services/configuration/browser/configurationService';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
