+++ src/vs/workbench/api/test/browser/mainThreadDiagnostics.test.ts
