+++ src/vs/workbench/api/test/browser/mainThreadDocumentContentProviders.test.ts
@@ -10,7 +10,7 @@ import { createTextModel } from 'vs/editor/test/common/testTextModel';
 import { mock } from 'vs/base/test/common/mock';
 import { IModelService } from 'vs/editor/common/services/model';
 import { IEditorWorkerService } from 'vs/editor/common/services/editorWorker';
-import { TestRPCProtocol } from 'vs/workbench/test/browser/api/testRPCProtocol';
+import { TestRPCProtocol } from 'vs/workbench/api/test/common/testRPCProtocol';
 import { TextEdit } from 'vs/editor/common/languages';
 
 suite('MainThreadDocumentContentProviders', function () {
