+++ src/vs/workbench/api/test/browser/mainThreadDocuments.test.ts
