+++ src/vs/workbench/api/test/browser/mainThreadTreeViews.test.ts
